refactor FindNearestColor to palette
