+++ Core/Graphics/Palettes/Palette.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Numerics;
 using static Helion.Util.Assertion.Assert;
 
 namespace Helion.Graphics.Palettes;
@@ -19,6 +20,8 @@ public class Palette
     private static Palette? DefaultPalette;
 
     private readonly List<Color[]> layers;
+    private readonly Vector3[] m_paletteNormalized = new Vector3[256];
+    private readonly Dictionary<Color, byte> m_colorToIndex = [];
 
     public int Count => layers.Count;
     public Color[] DefaultLayer => layers[0];
@@ -28,6 +31,14 @@ public class Palette
     {
         layers = paletteLayers;
         Translation = translation;
+
+        var colors = paletteLayers[0];
+        for (int i = 0; i < colors.Length; i++)
+        {
+            var c = colors[i];
+            m_paletteNormalized[i] = new Vector3(c.R / 255f, c.G / 255f, c.B / 255f);
+            m_colorToIndex[c] = (byte)i;
+        }
     }
 
     /// <summary>
@@ -122,5 +133,28 @@ public class Palette
 
         DefaultPalette = palette;
         return palette;
+    }
+
+    public byte GetNearestColorIndex(Color color)
+    {
+        if (m_colorToIndex.TryGetValue(color, out var index))
+            return index;
+
+        byte bestIndex = 0;
+        var nearest = float.MaxValue;
+        var colorNormalized = new Vector3(color.R / 255f, color.G / 255f, color.B / 255f);
+        for (int i = 0; i < m_paletteNormalized.Length; i++)
+        {
+            var paletteColor = m_paletteNormalized[i];
+            var calc = Vector3.DistanceSquared(colorNormalized, paletteColor);
+            if (calc < nearest)
+            {
+                bestIndex = (byte)i;
+                nearest = calc;
+            }
+        }
+
+        m_colorToIndex[color] = bestIndex;
+        return bestIndex;
     }
 }
